PR COMMENTS + custom hook
- remove console.log and themeContext from all Components
- add custom hook for all async requests
