+++ src/Components/Button/Button.jsx
@@ -1,10 +1,9 @@
-import { ThemeContext } from '../../context/themeContext';
-import { useContext } from 'react';
+import { useSelector } from 'react-redux';
 
 import styles from './Button.module.scss';
 
 function Button({ children }) {
-  const { theme } = useContext(ThemeContext);
+  const theme = useSelector((state) => state.theme);
 
   return <button className={styles[`${theme}-theme--btn`]}>{children}</button>;
 }
