+++ src/Components/FeedbackPage/FeedbackForm/FeedbackForm.jsx
@@ -1,14 +1,15 @@
-import { useEffect, useState, useContext } from 'react';
+import { useEffect, useState } from 'react';
+import { useSelector } from 'react-redux';
+
 import { saveState } from '../../../DataService/localDataService';
 
 import { ErrorMessage } from '../../';
-import { ThemeContext } from '../../../context/themeContext';
 import { Button } from '../../';
 
 import styles from './FeedbackForm.module.scss';
 
 function FeedbackForm() {
-  const { theme } = useContext(ThemeContext);
+  const theme = useSelector((state) => state.theme);
 
   const initialValues = {
     username: '',
